CMake: Load context-sensitive help.
CMake 3 uses reStructuredText and Sphinx for documentation. Those
tools can generate Qt qch files, which makes it possible to load
the documentation in QtCreator and to enable context-sensitive
help when editing a cmake language file.

Add a simple parser to process commands. This may be extended in the
future to process variables and properties too.

Change-Id: Ic5a49a9ce828ad9a9e8ff4805b48302ee73bcc71