C++: support smart splitting of strings.
If 'enter' is pressed while the cursor is in the middle of a string, the string is ended at the
current cursor position, and a new string is started on the next line. This makes it very easy to
split a long string onto multiple lines.

In addition, Shift+Enter insert an escape in the string, to continue the string at the beginning of
next line.

A setting can be used to enable or disable this option.

Change-Id: Ia5f3c6989fc00d40d06bc4fe1182fe8b1318f565